Create PIAVPN
Quick and easy file for importing over and then you can use it to randomize the connection...and then well...connect. This is only for a macbook and for Private Internet Access vpn service. Your terminal has to have access to the 'piactl' command. So type that into your terminal first to make sure it works. Then you import over this file into whatever project your working on and you've got 3 options of setting up a random connection, or you can simply connect and it will default to auto..